Adding Spanish translation for the Launcher, English fixes
Greetings. As I have talked with @Ratiocinator, I have done a first Spanish translation for the Launcher app. The translation has been tested and it displays some issues that probably happen because multilanguage support wasn't fully checked for this release.

I am also adding fixes to the English texts:
 - Typo correction.
 - Unification of the term "Red Pyramid Thing" across the board.
 - Unification of the term "scenario" for Maria and James' stories, following the official texts.
 - Added "apartment" to the "Change closet spawn" option name to follow the information displayed in other settings.
 - Removed Title Uppercasing in a few entries (you will need to decide which casing is required for the option names, though. I just changed those options that were the minority).

Issues currently happening with the Launcher and non-English languages:
 - The description box is small, specially for the Fullscreen Image Scaling option.
 - The Defaults button is not scalable.

I do know that multilanguage support is a thing that might be for the future, but I just wanted to get things done when I had some time to do so.